requirements.txt
asgiref==3.2.10
automium==0.2.6
automium-web==0.1.1
click==7.1.2
Django==3.1.1
django-ckeditor==6.0.0
django-filter==1.1.0
django-js-asset==1.2.2
djangorestframework==3.12.1
djangorestframework-filters==0.11.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
MarkupSafe==1.1.1
Pillow==8.0.1
pytz==2020.1
sqlparse==0.3.1
views==0.3
Werkzeug==1.0.1
